+++ src/client/components/pages/LandingPage.tsx
@@ -1,13 +1,10 @@
-import { useGoogleLogin } from "@react-oauth/google";
-import { useQueryClient } from "@tanstack/react-query";
 import { Button, Flex, Typography, message } from "antd";
 import React from "react";
 import { useNavigate } from "react-router-dom";
 import styled from "styled-components";
-import { api } from "../../api";
 import { useAuth } from "../../hooks/useAuth";
 import { useCurrentUserProfile } from "../../hooks/useCurrentUserProfile";
-import { useSetupProfileModal } from "../../hooks/useSetupProfileModal";
+import { GoogleLoginButton } from "../GoogleLoginButton";
 
 export function LandingPage() {
   const { logoutUser } = useAuth();
@@ -55,42 +52,6 @@ export function LandingPage() {
         )}
       </Flex>
     </Container>
-  );
-}
-
-function GoogleLoginButton() {
-  const queryClient = useQueryClient();
-  const { openSetupProfileModal } = useSetupProfileModal();
-
-  const googleLogin = useGoogleLogin({
-    onSuccess: async (e) => {
-      const token = e.access_token;
-      let success = await api.authLoginWithGoogle(token);
-
-      if (!success) {
-        const profileInfo = await openSetupProfileModal();
-        success = await api.authSignupWithGoogle(token, profileInfo);
-      }
-
-      if (success) {
-        message.info("Logged in!");
-        await queryClient.refetchQueries({
-          queryKey: ["getCurrentUserProfile"],
-        });
-      }
-    },
-    onError: (e) => {
-      message.error(`Google auth failed (${e})`);
-    },
-  });
-
-  return (
-    <Button onClick={() => googleLogin()}>
-      <Flex gap={8} justify="center" align="center">
-        <img src="svg/google.svg" width={18} />
-        Continue with Google
-      </Flex>
-    </Button>
   );
 }
